added confirm prompts to delete requests

+++ client/src/services/CommentsService.js
@@ -1,12 +1,15 @@
 import { AppState } from "../AppState.js"
 import { Comment } from "../models/Comment.js"
 import { logger } from "../utils/Logger.js"
+import Pop from "../utils/Pop.js"
 import { api } from "./AxiosService.js"
 
-class CommentsService{
+class CommentsService {
     async deleteComment(id) {
+        const confirm = await Pop.confirm('Are you sure you want to delete this comment?')
+        if (!confirm) { return ('canceled') }
         await api.delete(`api/comments/${id}`)
-        const index = AppState.activeComments.findIndex((comment) =>  comment.id == id )
+        const index = AppState.activeComments.findIndex((comment) => comment.id == id)
         AppState.activeComments.splice(index, 1)
     }
 
@@ -16,9 +19,9 @@ class CommentsService{
         AppState.activeComments.unshift(comment)
     }
 
-    async getCommentsForEvent(eventId){
-        const response =  await api.get(`api/events/${eventId}/comments`)
-        const comments = response.data.map((commentData)=>new Comment(commentData))
+    async getCommentsForEvent(eventId) {
+        const response = await api.get(`api/events/${eventId}/comments`)
+        const comments = response.data.map((commentData) => new Comment(commentData))
         AppState.activeComments = comments.reverse()
         logger.log(comments)
     }
